requirements.txt
asgiref==3.2.5
attrs==19.3.0
Django==2.2
django-crispy-forms==1.9.0
django-filter==2.2.0
django-jsonforms==1.1.0
django-tables2==2.2.1
importlib-metadata==1.5.0
jsonschema==3.2.0
psycopg2-binary==2.8.4
pyrsistent==0.15.7
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
zipp==3.1.0
